projects: ad463x_fmcz: use the example infrastructure
Migrate ad463x_fmcz to the shared examples infrastructure (examples.mk):
replace the Makefile BASIC_EXAMPLE/IIO_EXAMPLE toggles with EXAMPLE := iio
+ include examples.mk, move the iio sources into
src/examples/iio/example.mk (IIOD=y), and move the xilinx-only
SPI_ENGINE_OFFLOAD_EXAMPLE define into src.mk (it keys on PLATFORM). The
project-wide AD463X_ID device-select stays in src.mk. Drop examples_src.mk
and platform_includes.h, shorten the example directories
(basic_example -> basic, iio_example -> iio), and slim all three platform
main.c files (stm32 keeps stm32_init, xilinx keeps the cache enables) to
defer to example_main(); the basic example owns its UART stdio init and
the iio example keeps iio.h/iio_app.h. Switch builds.json flags to
EXAMPLE= (keeping the xilinx hardware arrays).

Verified by dependency resolution on both platforms: all four (example,
platform) combos resolve their full SRCS/INCS with the correct device and
offload defines, and every referenced file exists.
