+++ projects/ad463x_fmcz/src/examples/iio/example.mk
@@ -1,0 +1,4 @@
+IIOD = y
+
+INCS += $(DRIVERS)/adc/ad463x/iio_ad463x.h
+SRCS += $(DRIVERS)/adc/ad463x/iio_ad463x.c
