+++ projects/ad463x_fmcz/src/platform/stm32/main.c
@@ -31,15 +31,11 @@
  * EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 *******************************************************************************/
 
-#include "platform_includes.h"
+#include "parameters.h"
 #include "common_data.h"
 #include "no_os_error.h"
 
-#ifdef BASIC_EXAMPLE
-#include "basic_example.h"
-#elif defined(IIO_EXAMPLE)
-#include "iio_example.h"
-#endif
+extern int example_main();
 
 /***************************************************************************//**
  * @brief Main function execution for STM32 platform.
@@ -48,25 +44,7 @@
 *******************************************************************************/
 int main()
 {
-	int ret = -EINVAL;
-
 	stm32_init();
 
-#ifdef BASIC_EXAMPLE
-	struct no_os_uart_desc *uart;
-
-	ret = no_os_uart_init(&uart, &ad463x_uart_ip);
-	if (ret)
-		return ret;
-
-	no_os_uart_stdio(uart);
-
-	ret = basic_example_main();
-#elif defined(IIO_EXAMPLE)
-	ret = iio_example_main();
-#else
-#error At least one example has to be selected using y value in Makefile.
-#endif
-
-	return ret;
+	return example_main();
 }
